Code and documentation cleanup.
Cleaned some wired conversions from QString to const *char, AqBanking
expected strings in utf-8 and the conversion from QString to utf-8 is
easier than the version used to convert to const *char.

Also the todo-text for the doxygen docu in abt_job_ctrl.cpp changed, so
that all text belongs to the todo.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@354 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0