Support for SEPA transfers.
SEPA transfers could now be created and send to the bank. This is
implemented but not tested yet!

Also the known recipients are now capable of IBAN and BIC fields.
If an known recipient is draged to the receiver field of a transfer
form, the needed information is automatically filled with the account-


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@359 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0