updated translation for use with the new TranslationChooser class.
The translations for TranslationChooser:LANGUAGE and
TranslationsChooser:LANG_VERSION were set.

The language switch at runtime work and also the translation file
selection from different places (including build in resources file).
Not all strings are updated at this time. 
The changeEvent for LanguageChange must be adjusted to translate all
strings that are not used in the ui file (hopefully only at
mainwindow.cpp)


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/branch/dynamic_translations@441 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0