Enhanced translation behaviour on first start.
Now the first start consider the system locale and only falls back to
the default language (German) if no translation file matches the
language code.
At first run the country code is respected, but when nothing exactly
matching is found, the first matching language code file is used.

This should have no effect if the user selected a language.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/branch/dynamic_translations@445 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0