Updated project file.
Merged changes from /branch/dynamic_translations/abtransfers.pro:r451 to
trunk and applied changes to the use of pkg-config from Micha Lenk.

The changes at /branch/dynamic_translations/abtransfers.pro were merged
and changed in different ways. I don't know how to use some of this
changes for this merge. Therefore only the changes of r451 were merged
and edited.

The abtransfers.pro file at /trunk must be merged with the the other
changes applied to /branch/dynamic-translations and /branch/abt-mkae
when the corresponding changes are moved to /trunk.

This would be done soon after the release of version 0.0.4.1


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@477 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0