All strings should now translateable at runtime. refs #11 in trac.
Except of already open transaction forms (transfers, standing orders
etc.) all strings are translated at the time of a language change at
runtime.

The translation of already open transaction forms will be left open,
because a redesign of the forms in future releases is planned and would
only add unnecessary work at the moment!
A information message is shown to inform about this exception.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/branch/dynamic_translations@502 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0