finished the work on the calculator. fixes #4 in trac.
The calculator next to the value input field now works.

You could click the button to get the calculator or simple type the
first value into the input field and then press e.g. +, - ... and the
calculator will popup with the value entered previous and the operation
already loaded. So just type the second value and at return press the
calculated value will be set.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@524 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0