added APP_VERSION to the TranslationChooser. refs #38 in trac.
The APP_VERSION must be translated to the version of AB-Transfers for
which the translation should be used. When the translated version number
does not match the version number of AB-Transfers a warning is shown
when switching to the translation.

The warning message is not displayed if the translation is from the
embedded resources file or the default language (German). In this cases
it should not happen that the translation is outdated and the user can't
fix this.

The warning message could be disabled by the user. When the warning is
disabled it is disabled until a newer version of AB-Transfers is used or
it could be enabled again by hand at the settings.ini file (The option
ShowWarnTrans_LANGUAGENAME-VERSION must be set to true).


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/branch/dynamic_translations@529 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0