added support for automated export. fixes #3 in trac.
At the settings the export plugin and profile could be selected. Every
transaction is exported by the selected profile to the supplied file
(depending on which transaction is supported by the profile).

The export file is overwritten with each newly executed jobs! If you
need to import them to another program this must be done after each
execution of the jobs from the outbox.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@584 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0