fixed a crash with newer gwenhywfar versions. fixes #46 in trac.
The deletion of the used GWEN_GUI now uses a workaround to probably set
the internal used GWEN_GUI reference in gwenhywfar to NULL. So that the
gwen logger functions do not try to use the no longer existing gui.

This was tested with the following versions of gwenhywfar and aqbanking
and with the GWEN_LOGLEVEL environment variable set and not set.

gwenhywfar: v4.3.3 and latest svn revision 2246 (v4.7.0beta)
aqbanking: v5.0.24 and latest svn revision 2383 (v5.0.31beta)


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@603 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0