fixed compilation with AqBanking >= 5.5.0
At AqBanking Version >= 5.5.0 the AB_Transaction_GetCreditorItendifier()
and AB_Transaction_GetMandateReference() functions are removed (as well
as the corresponding _Set...() functions).

The functions are declared in abtransfers but newer used yet.
So the function calls are simply replaced by the corresponding
AB_Transaction_GetCreditorSchemeId() and AB_Transaction_GetMandateId()
functions (as well as the _Set...() functions).

This should also work with previous versions of AqBanking (compilation
tested with version 5.4.3.0).


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@620 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0