implemented "Get SEPA Standing Orders" (AqBanking >= 5.5.0 required!)
This is a first test with the newly added job to retrieve the standing
orders with the usage of SEPA jobs (added at AqBanking >= 5.5.0).

Sending the job to the bank is supported in AqBanking, but the parsing
of the returned values is not done yet at AqBanking. Therefore this job
could only be send by abtransfers but there are no values that could be
parsed and so the standing orders are not updated!

At the moment the execution of addGetSepaStandingOrders() is commented
out at MainWindow::onActionStandingUpdateTriggered().

All newly added functions are only compiled when the used AqBanking
version is >= 5.5.0 therefore the code is also useable with older
versions of AqBanking without any modifications!


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@621 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0