compilation with Qt5
fixed various failures when compiling with Qt5.

AB-Transfers does not work when compiling with Qt5 but the failure
should be an issue when AqBanking/gwenhywfar were comilied against Qt4.

I was not able to compile gwenhywfar with Qt5 even when the configure
options for --with-qt4...= are all pointed to my Qt5.3 installation.

Perhaps in a further release of gwenhywfar and aqbanking the Qt5 support
would be added and then abtransfer would also be ready to compile againt
Qt5.

The compilation against Qt 4.8.3 still works and is useable.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@622 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0