refactored the abt_parser
The parser is now splitted into multiple functions which do only exactly
one job of the processing. Also the different parser for the history
data is now inlcuded in the normal parser.

_All_ data could now be parsed by abt_parser::parse_ctx() !

Also documented the functions in english and the debug output of the
parser could be disabled (default) or enabled.

The parser should work like the previous version, because the previous
parser was only splitted into different parts. Therefore no detailed
tests were performed! If something does not work as expected the parser
code should be reviewed!


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@623 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0