Fixed crash when the export menu gets closed. fixes #55 and refs #37 in trac.
When the export menu gets closed without selecting any entry the program
crashes. This could also be verified when a selection was made and the
export was successfull.

I dont know why this works for me some time ago, perhaps an issue with
Qt?

Now it works every time, tested with Qt 4.8.6


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/trunk@625 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0