first tests and results for the ecoding problem
Activated the pageWidgetTests and tested the different possibilitys to
read and write an utf-8 encoded string to the imexporter module of
aqbanking. This works now.

It seems the is another issue (maybe AqBanking related). If the
transaction objects of AqBanking are used to fill out local parts at the
imexporter module, the same wired encoding happens again.

It seems that an already utf-8 encoded text is encoded another time.


git-svn-id: https://schmufu.dyndns.org/svn/ab_transfers/branch/encoding_problems@630 459a2fdc-afd4-4fb9-a8a9-e939c3f66bf0